LayoutTests/platform/glib/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entry-after-detach-expected.txt


FAIL NavigationHistoryEntry properties after detach assert_equals: expected (object) null but got (string) "http://web-platform.test:8800/common/blank.html"
